+++ src/mcp/server.rs
@@ -74,7 +74,7 @@ impl DashMcpService {
     pub fn tool_router() -> ToolRouter<Self> {
         ToolRouter::new()
             .with_async_tool::<tools::network::NetworkTool>()
-            .with_async_tool::<tools::network::ListWalletsTool>()
+            .with_async_tool::<tools::wallet::ListWalletsTool>()
             .with_async_tool::<tools::wallet::GenerateReceiveAddress>()
             .with_async_tool::<tools::wallet::WalletBalancesQuery>()
             .with_async_tool::<tools::wallet::FetchPlatformBalances>()
