+++ src/mcp/tools/network.rs
@@ -1,4 +1,4 @@
-//! Network and wallet-listing MCP tools.
+//! Network MCP tools.
 
 use std::borrow::Cow;
 
@@ -8,9 +8,8 @@ use rmcp::schemars;
 use serde::Serialize;
 
 use crate::mcp::error::McpToolError;
-use crate::mcp::resolve;
 use crate::mcp::server::{DashMcpService, collect_available, network_display_name};
-use crate::mcp::tools::{EmptyParams, NetworkParams};
+use crate::mcp::tools::EmptyParams;
 
 // ---------------------------------------------------------------------------
 // NetworkTool
@@ -42,10 +41,6 @@ impl ToolBase for NetworkTool {
         )
     }
 
-    fn input_schema() -> Option<std::sync::Arc<rmcp::model::JsonObject>> {
-        None
-    }
-
     fn annotations() -> Option<ToolAnnotations> {
         Some(ToolAnnotations::default().read_only(true).open_world(true))
     }
@@ -70,66 +65,5 @@ impl AsyncTool<DashMcpService> for NetworkTool {
             .collect();
 
         Ok(NetworkOutput { active, available })
-    }
-}
-
-// ---------------------------------------------------------------------------
-// ListWalletsTool
-// ---------------------------------------------------------------------------
-
-/// List wallet names currently loaded in the application.
-pub struct ListWalletsTool;
-
-#[derive(Serialize, schemars::JsonSchema)]
-pub struct WalletEntry {
-    seed_hash: String,
-    alias: Option<String>,
-}
-
-#[derive(Serialize, schemars::JsonSchema)]
-pub struct ListWalletsOutput {
-    wallets: Vec<WalletEntry>,
-}
-
-impl ToolBase for ListWalletsTool {
-    type Parameter = NetworkParams;
-    type Output = ListWalletsOutput;
-    type Error = McpToolError;
-
-    fn name() -> Cow<'static, str> {
-        "core_wallets_list".into()
-    }
-
-    fn description() -> Option<Cow<'static, str>> {
-        Some("List wallet names currently loaded in the application".into())
-    }
-
-    fn annotations() -> Option<ToolAnnotations> {
-        Some(ToolAnnotations::default().read_only(true).open_world(false))
-    }
-}
-
-impl AsyncTool<DashMcpService> for ListWalletsTool {
-    async fn invoke(
-        service: &DashMcpService,
-        param: NetworkParams,
-    ) -> Result<ListWalletsOutput, McpToolError> {
-        let ctx = service
-            .ctx()
-            .await
-            .map_err(|e| McpToolError::Internal(e.to_string()))?;
-        resolve::verify_network(&ctx, param.network.as_deref())?;
-        let wallets = ctx.wallets.read().unwrap_or_else(|e| e.into_inner());
-        let entries: Vec<WalletEntry> = wallets
-            .iter()
-            .map(|(hash, wallet_arc)| {
-                let wallet = wallet_arc.read().unwrap_or_else(|e| e.into_inner());
-                WalletEntry {
-                    seed_hash: hex::encode(hash),
-                    alias: wallet.alias.clone(),
-                }
-            })
-            .collect();
-        Ok(ListWalletsOutput { wallets: entries })
     }
 }
